Flatten Rollout Manager and Session Manager into an Execution Manager
