+++ reward_kit/mcp/execution/__init__.py
@@ -5,10 +5,10 @@ This module handles policy execution, tool calling, and rollout coordination.
 """
 
 from .policy import FireworksPolicy, LLMBasePolicy
-from .rollout import RolloutManager
+from .manager import ExecutionManager
 
 __all__ = [
     "LLMBasePolicy",
     "FireworksPolicy",
-    "RolloutManager",
+    "ExecutionManager",
 ]
