+++ reward_kit/mcp/session/__init__.py
@@ -4,9 +4,8 @@ MCP Session Management
 This module handles session management and vector environment operations.
 """
 
-from .manager import GeneralMCPVectorEnv, SessionManager
+from .manager import GeneralMCPVectorEnv
 
 __all__ = [
-    "SessionManager",
     "GeneralMCPVectorEnv",
 ]
